+++ monty/main.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
-from monty.model import *
 from monty.input import UnchangedInputFunction
+from monty.model import *
+
+tf.enable_eager_execution()
 
 if __name__ == '__main__':
     estimator = AutoEncoder(learning_rate=0.001, model_dir="out", num_latent_variables=100)
